Build Hang Resolved
After much debugging it was found that the `writeCommitHashToCoreJSON` gulp task was the source of the build hanging. The file `add` event was never triggering and the callback was never fired.

Because these tasks are run in series, I've altered this task to simply execute with the assumption that the file is already there and doesn't need to be watched for. This should work as long as the tasks are never run in parallel.