To splice, or not to... No. No splicing.
Changes the method in which we remove insurance packages from the profile. We were iterating over them using a `for` in reverse order and then removing them with a splice and the current index. On paper this should work, but funny things were happening... Sometimes. I've created a new method that removes packages from a profile by matching against the systemData's date, time, and location. This should give us a very specific insurance package (as we don't have an ID to use). In addition to this we're fetching a new instance of the profile to edit, instead of modifying packages that are being iterated over as they're being iterated over. Hopefully this straightforward approach leads to less "funny things" happening.

Improves some debugging messages.