Updated Project to use Biome for code formatting.
Details on prettier can be found here:
https://biomejs.dev/

VSC plugin:
https://marketplace.visualstudio.com/items?itemName=biomejs.biome

Before this, we were using both ESLint and Biome for linting. This commit removes ESLint and uses just Biome for linting. Additionally, it enables Biome formatting. When a file is saved (and you have the VSCode plugin is installed) it will format the file based on Biome's opinionated formatting rules. This speeds up code reviews as code *style* can now be ignored.