Fix being unable to examine items on the Scav sell screen
A previous commit added the Encyclopedia as an empty object for scavs. However this caused an unintended side effect of not being able to examine items at the post-raid screen for scavs.

This causes an issue where users are unable to sell equipment they have equipped on their scav that their PMC had already examined but their scav did not examine in-raid.