Updates Package Versioning
Changes versioning of the remaining packages (that do not require being pinned to a specific version) to allow for upgrades to hot-fixes within their current minor version.

For example, we'll allow an install of atomically `v1.7.x` now, instead of pinning to `v1.7.0`.

This should make upgrades a little easier in the future.