Prettier SWC Conflict
This change formats the `ModCompilerService` to match the new formatting rules and also adds it to the Prettier ignore file. When it's run automatically, Prittier interferes with a SWC qwerk with the following string:

"//# sourceMappingURL\=module.js.map"

The backslash before the equal is removed, as it's deemed unnecessary, however we're relying on it being there for SWC source mapping to function correctly.