added handler to prevent server from hanging infinitely
After reducing the list of possible items to fulfill the new budget constraint, it may happen that all of the remaining items have already been tried as target. 

Example: 
randomNumbersUsed = [2, 1, 0, 3], itemSelection.length = 4
The server in that case was stuck on trying to find a new number lower than itemSelection.length, but not used yet. Such number doesn't exist, obviously.

It can easily be reproduced by increasing the number of daily/weekly quests to generate to 100.