Adds DatabaseDecompressionUtil Class
This class is responsible for decompressing database archives. It's initialized on server start, but the decompression mechanism only fires when the target database directory is empty. Adds roughly 1.5 seconds to initial server start (in my limited testing).

The gulp build script has been updated to ignore the location database files.

Needs to be tested.