Adds decompressArchives Gulp task
This gulp task runs before any of the database assets are copied over. It ensures that all database archives are decompressed into their respective directories. This means that the archive has priority when building. All files within the target database directories will be removed and replaced with the contents of the database archives.

Uses the same two dev dependancies as the NPM 'database' scripts.