Workflow: Manually Checkout LFS Files
The `actions/checkout@v4` action does not support `.lfsconfig` files, which means that our custom LFS server can't be used with this action. The work around is to disable the LFS operation in the action then manually pull the LFS files in a step after the normal repo is pulled down.