Resolves error, adds readline package
I was getting an error about `cursorTo` not existing on `stdout`. This adds the node readline package to adjust the CLI cursor position.

This works in all of the terminals that I could test on, for the exception of the VSCode built-in terminal in which some escape codes are visible on the beginning of the line. Doesn't bother me, but if you can figure it out then please do.