Development Strict Mode
- Enables TS strict mode
- Adds script to check types
- Adds workflow to run check-type script
- Updates the code-checking workflows to share the same set-up job

This updates the `tsconfig.json` option to enable strict mode. *However*, we use TSX for development and SWC for release, which do not type-check, so this option only gives additional linting/visual-feedback in VSCode.

Additionally, I've added a NPM script `npm run lint:types` and a GitHub workflow that runs it.